Migrating to OpenGL 4.5

+++ boar-main/src/main/kotlin/net/spartanb312/boar/utils/math/MathUtils.kt
@@ -95,7 +95,7 @@ object MathUtils {
     }
 
     @JvmStatic
-    inline fun calcSegments(segmentsIn: Int, radius: Double, range: Float): Int {
+    inline fun calcSegments(segmentsIn: Int, radius: Float, range: Float): Int {
         if (segmentsIn != -0) return segmentsIn
         val segments = radius * 0.5 * PI * (range / 360.0)
         return max(segments.roundToInt(), 16)
